[#12] chore: ground, matchLog, scoreBoard 디렉토리 그룹핑
