+++ FE/client/src/components/match/Match.jsx
@@ -2,9 +2,9 @@ import React from 'react'
 import useSound from '../../utils/useSound';
 import { useParams } from 'react-router-dom';
 import styled from 'styled-components';
-import ScoreBoard from './ScoreBoard';
-import MatchLog from './MatchLog';
-import Ground from './Ground';
+import ScoreBoard from './scoreBoard/ScoreBoard';
+import MatchLog from './matchLog/MatchLog';
+import Ground from './ground/Ground';
 import MatchBGM from '../../audios/MatchBGM.mp3';
 
 const MatchWrap = styled.div`
